Use Derivation only on Scala 2.11+
Also extract macro hacks to a separate file for easier maintenance in the future.
