+++ core/src/test/scala-2.11+/pureconfig/DerivationSuite.scala
